.openclaw/workspace-biology/question_bank/requirements.txt
streamlit>=1.28
chromadb>=0.4
python-docx>=0.8
openpyxl>=3.1
sentence-transformers>=2.2
